build improvement

+++ bin/download.sh
@@ -9,47 +9,62 @@ export NVM_DIR="$HOME/.nvm"
 echo 'nvm:' $(nvm -v)
 nvm use > /dev/null 2>&1
 
-# cleanup
-rm -f -R ./download > /dev/null 2>&1
-
 # keccak
-mkdir -p ./download/keccak > /dev/null 2>&1
-echo '*** browserify keccak ***'
-yarn browserify \
-  node_modules/keccak/js.js \
-  -s createKeccakHash \
-  -o download/keccak/keccak.umd.js || exit 1
-yarn terser \
-  download/keccak/keccak.umd.js \
-  -c toplevel,sequences=false -m \
-  -o download/keccak/keccak.umd.min.js || exit 1
+if [ ! -e ./download/keccak ]
+then
+  mkdir -p ./download/keccak > /dev/null 2>&1
+  echo '*** browserify keccak ***'
+  yarn browserify \
+    node_modules/keccak/js.js \
+    -s createKeccakHash \
+    -o download/keccak/keccak.umd.js || exit 1
+  yarn terser \
+    download/keccak/keccak.umd.js \
+    -c toplevel,sequences=false -m \
+    -o download/keccak/keccak.umd.min.js || exit 1
+fi
 
 # down
 cd download
 
 # @metamask/detect-provider
-mkdir detect-provider > /dev/null 2>&1
-wget https://cdn.jsdelivr.net/npm/@metamask/detect-provider@1.2.0/dist/detect-provider.min.js -O ./detect-provider/detect-provider.min.js || exit 1
+if [ ! -e ./detect-provider ]
+then
+  mkdir detect-provider > /dev/null 2>&1
+  wget https://cdn.jsdelivr.net/npm/@metamask/detect-provider@1.2.0/dist/detect-provider.min.js -O ./detect-provider/detect-provider.min.js || exit 1
+fi
 
 # loglevel
-mkdir loglevel > /dev/null 2>&1
-wget https://cdn.jsdelivr.net/npm/loglevel@1.7.1/dist/loglevel.min.js -O ./loglevel/loglevel.min.js || exit 1
+if [ ! -e ./loglevel ]
+then
+  mkdir loglevel > /dev/null 2>&1
+  wget https://cdn.jsdelivr.net/npm/loglevel@1.7.1/dist/loglevel.min.js -O ./loglevel/loglevel.min.js || exit 1
+fi
 
 # pako
-mkdir pako > /dev/null 2>&1
-wget https://cdn.jsdelivr.net/npm/pako@2.0.3/dist/pako.min.js -O ./pako/pako.min.js || exit 1
+if [ ! -e ./pako ]
+then
+  mkdir pako > /dev/null 2>&1
+  wget https://cdn.jsdelivr.net/npm/pako@2.0.3/dist/pako.min.js -O ./pako/pako.min.js || exit 1
+fi
 
 # tw5-locator
-git clone https://github.com/bimlas/tw5-locator || exit 1
-cd tw5-locator
-git checkout tags/v2.0.1 || exit 1
-cd ..
+if [ ! -e ./tw5-locator ]
+then
+  git clone https://github.com/bimlas/tw5-locator || exit 1
+  cd tw5-locator
+  git checkout tags/v2.0.1 || exit 1
+  cd ..
+fi
 
 # tw5-relink
-git clone https://github.com/flibbles/tw5-relink || exit 1
-cd tw5-relink
-git checkout tags/v1.10.3 || exit 1
-cd ..
+if [ ! -e ./tw5-relink ]
+then
+  git clone https://github.com/flibbles/tw5-relink || exit 1
+  cd tw5-relink
+  git checkout tags/v1.10.3 || exit 1
+  cd ..
+fi
 
 # up
 cd ..
